+++ internal/driver/mobile/file_desktop.go
@@ -25,6 +25,11 @@ func nativeFileOpen(*fileOpen) (io.ReadCloser, error) {
 	return nil, nil
 }
 
+func nativeFileOpenSeeker(*fileOpen) (io.ReadSeekCloser, error) {
+	// no-op as we use the internal FileRepository
+	return nil, nil
+}
+
 func nativeFileSave(*fileSave, bool) (io.WriteCloser, error) {
 	// no-op as we use the internal FileRepository
 	return nil, nil